revise funnel for async
Old freq deduplication doesn't work for sending candidates as they arrive, so removed. Also now know that sync score is not a good indicator of decodeability, so deduplicating when this is the only metric is bad.

Also revise metric file output, timed log formatting, and move prioritisation of rxfreq into cycle manager (during the one candidate sort (which is repeated each hop to catch new candidates as they arrive).